+++ mobile-users/android/app/src/main/java/com/purnazen/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.wellness
+package com.purnazen
 
 import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
@@ -12,7 +12,7 @@ class MainActivity : ReactActivity() {
    * Returns the name of the main component registered from JavaScript. This is used to schedule
    * rendering of the component.
    */
-  override fun getMainComponentName(): String = "wellness"
+  override fun getMainComponentName(): String = "purnazen"
 
   /**
    * Returns the instance of the [ReactActivityDelegate]. We use [DefaultReactActivityDelegate]